requirements.txt
aiogram~=2.12.1
environs~=9.3.2
python-dotenv~=0.17.0
asgiref~=3.2.10
geopy~=2.1.0
Django==3.1
fake-useragent
psycopg2-binary
openpyxl~=3.0.8
djangorestframework==3.12.4
emoji==1.5.0
requests==2.26.0
apscheduler==3.8.0
django-simple-history==3.0.0
aspose-cells==21.11.0
pathlib~=1.0.1
easyocr~=1.4.1
pytesseract~=0.3.8
Pillow~=8.2.0